+++ projects/core/src/remult3/RepositoryImplementation.ts
@@ -402,6 +402,7 @@ export class RepositoryImplementation<entityType>
     where: EntityFilter<entityType>,
     set: Partial<MembersOnly<entityType>>,
   ): Promise<number> {
+    Filter.throwErrorIfFilterIsEmpty(where, 'updateMany')
     if (this._dataProvider.isProxy) {
       return (this._edp as any as ProxyEntityDataProvider).updateMany(
         await this._translateWhereToFilter(where),
@@ -880,6 +881,7 @@ export class RepositoryImplementation<entityType>
     return this._edp.count(await this._translateWhereToFilter(where))
   }
   async deleteMany(where: EntityFilter<entityType>): Promise<number> {
+    Filter.throwErrorIfFilterIsEmpty(where, 'deleteMany')
     if (this._dataProvider.isProxy) {
       return (this._edp as any as ProxyEntityDataProvider).deleteMany(
         await this._translateWhereToFilter(where),