+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 - Added support for migrations, see [Migrations](https://remult.dev/docs/migrations.html)
 - Added an `error` hook to `RemultServerOptions` that is called whenever whenever there is an error in the API lifecycle. See [RemultServerOptions](https://remult.dev/docs/ref_remultserveroptions.html#error)
+- Added an exception when calling `updateMany` or `dleleteMany` without a filter - to protect against accidential deleting / updating all data
+- Added updateMany and deleteMany to OpenAPI (swagger)
 - Changed postgres schema builder to use `timestamptz` instead of `timestamp`
 - `KnexDataProvider` now supports all `execute` and `createCommand` and can be used with any `SqlDatabase` functionality
 - **Breaking Change** changed the signature of `getDb` to recieve `DataProvider` as parameter instead of `Remult`